Move prover code around so it's cleaner.
